Introduce new ChunkRenderer class for sharing common programs, arrays etc.
